time.NewTicker

+++ test/base/15.atomic/main.go
@@ -11,7 +11,7 @@ func main() {
 
 	// We'll use an unsigned integer to represent our
 	// (always-positive) counter.
-	var ops uint64 = 0
+	var ops uint32 = 0
 
 	// To simulate concurrent updates, we'll start 50
 	// goroutines that each increment the counter about
